Prepared to change disordered enum translation

+++ repository/src/main/java/org/kofi/creditex/repository/UserRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.querydsl.QueryDslPredicateExecutor;
 
 public interface UserRepository extends JpaRepository<User, Long>, QueryDslPredicateExecutor<User> {
     User findByUsername(String username);
-    User findByUserData_FirstAndUserData_LastAndUserData_PatronymicAndUserData_PassportSeriesAndUserData_PassportNumber(String userData_First, String userData_Last, String userData_Patronymic, String userData_Series, long userData_Number);
+    User findByUserData_FirstAndUserData_LastAndUserData_PatronymicAndUserData_PassportSeriesAndUserData_PassportNumber(String userData_First, String userData_Last, String userData_Patronymic, String userData_Series, int userData_Number);
 }
